{"schema":"decodex/commit/1","summary":"Modularize operator status rendering","authority":"XY-1115"}
