+++ apps/decodex/src/orchestrator/status_render.rs
@@ -1,4 +1,32 @@
-fn render_operator_status(snapshot: &OperatorStatusSnapshot) -> String {
+use super::{
+	ChildAgentActivityBucket, ChildAgentActivitySummary, CodexAccountActivitySummary,
+	EXECUTION_LIVENESS_PROCESS_IDENTITY_MISMATCH, OperatorContinuationRecoveryStatus,
+	OperatorGitHubCliAuthority, OperatorHistoryLaneStatus, OperatorHistoryLedgerOutcome,
+	OperatorLaneLifecycleMetrics, OperatorLoopStatus, OperatorPhaseAcceptanceStatus,
+	OperatorQueuedIssueStatus, OperatorRunControlCapability, OperatorRunStatus,
+	OperatorSnapshotWarningDetail, OperatorStatusSnapshot, OperatorWorktreeStatus,
+	ProtocolActivityEventSummary, ProtocolActivitySummary,
+	QUEUE_REASON_LINEAR_ACTIVE_LABEL_PRESENT, ServiceConfig, format_optional_i64,
+	format_optional_unix_timestamp, format_status_no_eligible_issue_hint,
+	operator_protocol_activity_detail_is_public, operator_run_counts_as_current_lane,
+	operator_run_counts_as_running, operator_run_has_recent_app_server_execution,
+	project_attention_count, project_history_only_attention_count,
+	queued_candidate_counts_as_waiting_intake, render_private_evidence_reference,
+};
+
+mod activity;
+mod run_rows;
+
+use activity::{
+	render_account_summary, render_accounts_summary, render_child_agent_activity_summary,
+	render_child_agent_context_pressure, render_control_capability_summary,
+	render_loop_architecture_recovery_summary, render_loop_autonomy_signals_summary,
+	render_loop_boundary_summary, render_loop_review_summary, render_loop_status_summary,
+	render_protocol_activity_summary,
+};
+use run_rows::{append_rendered_history_lane, append_rendered_run};
+
+pub(crate) fn render_operator_status(snapshot: &OperatorStatusSnapshot) -> String {
 	let session_history_attempt_count =
 		snapshot.history_lanes.iter().map(|lane| lane.attempt_count).sum::<usize>();
 	let hides_current_lanes = session_history_attempt_count < snapshot.recent_runs.len();
@@ -311,7 +339,7 @@ fn format_warning_detail(detail: &OperatorSnapshotWarningDetail) -> String {
 	parts.join(" ")
 }
 
-fn render_queue_explain(
+pub(crate) fn render_queue_explain(
 	config: &ServiceConfig,
 	queued_candidates: &[OperatorQueuedIssueStatus],
 ) -> String {
@@ -374,7 +402,7 @@ fn rendered_backlog_queue_groups(
 	(stale_closed_queue_labels, backlog_candidates)
 }
 
-fn rendered_recovery_worktrees(
+pub(in crate::orchestrator) fn rendered_recovery_worktrees(
 	snapshot: &OperatorStatusSnapshot,
 ) -> Vec<(&str, &OperatorWorktreeStatus)> {
 	let mut rendered_worktrees = snapshot
@@ -598,770 +626,4 @@ fn rendered_worktree_role_rank(role: &str) -> u8 {
 		"post_review_lane" => 1,
 		_ => 2,
 	}
-}
-
-fn render_child_agent_activity_summary(summary: Option<&ChildAgentActivitySummary>) -> String {
-	let Some(summary) = summary else {
-		return String::from("none");
-	};
-	let current = match (&summary.current_bucket, summary.current_elapsed_seconds) {
-		(Some(bucket), Some(seconds)) => format!("{bucket} {}", format_seconds_compact(seconds)),
-		(Some(bucket), None) => bucket.clone(),
-		(None, _) => String::from("none"),
-	};
-	let buckets = render_child_agent_bucket_distribution(&summary.buckets);
-
-	format!(
-		"current={current}; wall={}; buckets={}; tool_calls={}",
-		format_seconds_compact(summary.wall_seconds),
-		buckets,
-		summary.tool_call_count
-	)
-}
-
-fn render_protocol_activity_summary(summary: Option<&ProtocolActivitySummary>) -> String {
-	let Some(summary) = summary else {
-		return String::from("none");
-	};
-	let turn = summary.turn_status.as_deref().unwrap_or("none");
-	let wait = summary.waiting_reason.as_deref().unwrap_or("none");
-	let rate_limit = summary.rate_limit_status.as_deref().unwrap_or("none");
-	let recent = if summary.recent_events.is_empty() {
-		String::from("none")
-	} else {
-		summary
-			.recent_events
-			.iter()
-			.rev()
-			.take(5)
-			.map(render_protocol_activity_event_summary)
-			.collect::<Vec<_>>()
-			.join(", ")
-	};
-
-	format!("turn={turn}; waiting={wait}; rate_limit={rate_limit}; recent={recent}")
-}
-
-fn render_protocol_activity_event_summary(event: &ProtocolActivityEventSummary) -> String {
-	event.detail.as_ref().map_or_else(
-		|| event.event_type.clone(),
-		|detail| format!("{}:{}", event.event_type, render_protocol_activity_detail(detail)),
-	)
-}
-
-fn render_protocol_activity_detail(detail: &str) -> &str {
-	if operator_protocol_activity_detail_is_public(detail) {
-		detail
-	} else {
-		"redacted_sensitive_detail"
-	}
-}
-
-fn render_loop_status_summary(status: Option<&OperatorLoopStatus>) -> String {
-	let Some(status) = status else {
-		return String::from("none");
-	};
-	let next_action = status.next_action.as_deref().unwrap_or("none");
-	let autonomy_objective = status
-		.autonomy_objective
-		.as_ref()
-		.map(|objective| objective.source_ref.as_str())
-		.unwrap_or("none");
-	let autonomy_report =
-		status.autonomy_report.as_ref().map(|report| report.authority.as_str()).unwrap_or("none");
-
-	format!(
-		"{}; review_level={}; autonomy={}; autonomy_objective={autonomy_objective}; autonomy_signals={}; autonomy_proposals={}; report={autonomy_report}; next_action={next_action}",
-		status.summary,
-		status.review_level,
-		status.autonomy,
-		status.autonomy_signals.len(),
-		status.autonomy_proposals.len()
-	)
-}
-
-fn render_loop_autonomy_signals_summary(status: Option<&OperatorLoopStatus>) -> String {
-	let Some(status) = status else {
-		return String::from("none");
-	};
-
-	if status.autonomy_signals.is_empty() {
-		return String::from("none");
-	}
-
-	status
-		.autonomy_signals
-		.iter()
-		.map(|signal| {
-			format!(
-				"{}:{}@v{} freshness={} confidence={} privacy={} sources={} completeness={} gaps={} contradictions={}",
-				signal.kind,
-				signal.objective_id,
-				signal.objective_version,
-				signal.freshness,
-				signal.confidence,
-				signal.privacy,
-				signal.source_refs.len(),
-				signal.completeness,
-				signal.gaps.len(),
-				signal.contradictions.len()
-			)
-		})
-		.collect::<Vec<_>>()
-		.join(";")
-}
-
-fn render_loop_review_summary(status: Option<&OperatorLoopStatus>) -> String {
-	let Some(review) = status.and_then(|status| status.review.as_ref()) else {
-		return String::from("none");
-	};
-	let checkpoint = review.checkpoint.as_ref().map_or_else(
-		|| String::from("checkpoint=none"),
-		|checkpoint| {
-			format!(
-				"checkpoint=head:{} round:{} review_class:{} risk_class:{} compact_eligible:{} fallback:{} updated:{}",
-				checkpoint.head_sha,
-				checkpoint.round,
-				checkpoint.review_class.as_deref().unwrap_or("none"),
-				checkpoint.risk_class.as_deref().unwrap_or("none"),
-				checkpoint
-					.compact_eligible
-					.map_or("none", |eligible| if eligible { "true" } else { "false" }),
-				checkpoint.fallback_reason.as_deref().unwrap_or("none"),
-				checkpoint.updated_at
-			)
-		},
-	);
-
-	format!("phase={} status={} {checkpoint}", review.phase, review.status)
-}
-
-fn render_loop_architecture_recovery_summary(status: Option<&OperatorLoopStatus>) -> String {
-	let Some(recovery) = status.and_then(|status| status.architecture_recovery.as_ref()) else {
-		return String::from("none");
-	};
-	let budget = recovery.budget.as_ref().map_or_else(
-		|| String::from("none"),
-		|budget| format!("{}/{}", budget.attempt, budget.max_attempts),
-	);
-
-	format!(
-		"status={} reason={} guardrail={} boundary={} policy={} enhanced_evidence={} blocks_landing={} budget={} next_action={}",
-		recovery.status,
-		recovery.reason_code,
-		recovery.guardrail_reason.as_deref().unwrap_or("none"),
-		recovery.boundary_disposition.as_deref().unwrap_or("none"),
-		recovery.boundary_policy_decision.as_deref().unwrap_or("none"),
-		recovery.requires_enhanced_evidence,
-		recovery.blocks_landing,
-		budget,
-		recovery.next_action
-	)
-}
-
-fn render_loop_boundary_summary(status: Option<&OperatorLoopStatus>) -> String {
-	let Some(boundary) = status.and_then(|status| status.boundary.as_ref()) else {
-		return String::from("none");
-	};
-
-	format!(
-		"disposition={} policy={} enhanced_evidence={} blocks_landing={} reason={} attempted_recovery={} changed_surfaces={} improvement_signals={}",
-		boundary.disposition,
-		boundary.policy_decision,
-		boundary.requires_enhanced_evidence,
-		boundary.blocks_landing,
-		boundary.reason.as_deref().unwrap_or("none"),
-		boundary.attempted_recovery_reason.as_deref().unwrap_or("none"),
-		boundary.changed_surface_count,
-		boundary.improvement_signal_count
-	)
-}
-
-fn render_control_capability_summary(capability: Option<&OperatorRunControlCapability>) -> String {
-	let Some(capability) = capability else {
-		return String::from("none");
-	};
-	let thread_id = capability.thread_id.as_deref().unwrap_or("none");
-	let turn_id = capability.turn_id.as_deref().unwrap_or("none");
-
-	format!(
-		"status={}; transport={}; channel={}; thread_id={thread_id}; turn_id={turn_id}",
-		capability.status, capability.transport, capability.channel_path
-	)
-}
-
-fn render_account_summary(summary: Option<&CodexAccountActivitySummary>) -> String {
-	let Some(summary) = summary else {
-		return String::from("none");
-	};
-	let plan = summary.plan_type.as_deref().unwrap_or("unknown");
-	let reached = summary.rate_limit_reached_type.as_deref().unwrap_or("none");
-	let credits = render_codex_account_credits(summary);
-	let token_status = render_codex_account_token_status(&summary.refresh_status);
-	let primary = render_codex_account_window(
-		summary.primary_window_seconds,
-		summary.primary_remaining_percent,
-		summary.primary_resets_at_unix_epoch,
-	);
-	let secondary = render_codex_account_window(
-		summary.secondary_window_seconds,
-		summary.secondary_remaining_percent,
-		summary.secondary_resets_at_unix_epoch,
-	);
-
-	format!(
-		"account={}; plan={plan}; status={}; token={token_status}; primary={primary}; secondary={secondary}; credits={credits}; reached={reached}",
-		summary.account_fingerprint, summary.status,
-	)
-}
-
-fn render_accounts_summary(accounts: &[CodexAccountActivitySummary]) -> String {
-	if accounts.is_empty() {
-		return String::from("none");
-	}
-
-	accounts
-		.iter()
-		.map(|summary| render_account_summary(Some(summary)))
-		.collect::<Vec<_>>()
-		.join(" | ")
-}
-
-fn render_codex_account_window(
-	window_seconds: Option<i64>,
-	remaining_percent: Option<i64>,
-	resets_at_unix_epoch: Option<i64>,
-) -> String {
-	let label = window_seconds.map(codex_window_label).unwrap_or_else(|| String::from("window"));
-	let remaining =
-		remaining_percent.map_or_else(|| String::from("unknown"), |value| format!("{value}%"));
-	let reset = format_optional_unix_timestamp(resets_at_unix_epoch)
-		.unwrap_or_else(|| String::from("unknown"));
-
-	format!("{label} remaining={remaining} reset={reset}")
-}
-
-fn render_codex_account_credits(summary: &CodexAccountActivitySummary) -> String {
-	if summary.credits_unlimited == Some(true) {
-		return String::from("unlimited");
-	}
-
-	match (summary.credits_has_credits, summary.credits_balance.as_deref()) {
-		(Some(false), Some(balance)) => format!("depleted balance={balance}"),
-		(Some(false), None) => String::from("depleted"),
-		(_, Some(balance)) => format!("balance={balance}"),
-		(Some(true), None) => String::from("available"),
-		(None, None) => String::from("unknown"),
-	}
-}
-
-fn render_codex_account_token_status(refresh_status: &str) -> &'static str {
-	match refresh_status {
-		"not_needed" | "none" => "ok",
-		"succeeded" | "refreshed" => "refreshed",
-		"failed" => "refresh_failed",
-		_ => "unknown",
-	}
-}
-
-fn codex_window_label(window_seconds: i64) -> String {
-	match window_seconds {
-		18_000 => String::from("5h"),
-		604_800 => String::from("7d"),
-		seconds => format_seconds_compact(seconds),
-	}
-}
-
-fn render_child_agent_context_pressure(summary: Option<&ChildAgentActivitySummary>) -> String {
-	let Some(summary) = summary else {
-		return String::from("none");
-	};
-	let current_input = summary
-		.input_tokens_current
-		.map(format_count_compact)
-		.unwrap_or_else(|| String::from("none"));
-	let max_input =
-		summary.input_tokens_max.map(format_count_compact).unwrap_or_else(|| String::from("none"));
-	let max_input_relation = match (summary.input_tokens_current, summary.input_tokens_max) {
-		(Some(current), Some(max)) if current == max => " (same as current)",
-		_ => "",
-	};
-	let largest_output = summary
-		.largest_tool_output_bytes
-		.map(format_bytes_compact)
-		.unwrap_or_else(|| String::from("none"));
-	let largest_tool = summary.largest_tool_output_tool.as_deref().unwrap_or("none");
-	let warnings = if summary.large_output_warnings.is_empty() {
-		String::from("none")
-	} else {
-		summary.large_output_warnings.join(" | ")
-	};
-
-	format!(
-		"input=current_window {current_input}, peak_window {max_input}{max_input_relation}, cumulative_input {}; output_tokens={}; largest_output={largest_output} by {largest_tool}; warnings={warnings}",
-		format_count_compact(summary.input_tokens_cumulative),
-		format_count_compact(summary.output_tokens_cumulative)
-	)
-}
-
-fn render_child_agent_bucket_distribution(buckets: &[ChildAgentActivityBucket]) -> String {
-	if buckets.is_empty() {
-		return String::from("none");
-	}
-
-	let mut buckets = buckets.iter().collect::<Vec<_>>();
-
-	buckets.sort_by(|left, right| {
-		right
-			.wall_seconds
-			.cmp(&left.wall_seconds)
-			.then_with(|| right.event_count.cmp(&left.event_count))
-			.then_with(|| left.name.cmp(&right.name))
-	});
-
-	buckets
-		.into_iter()
-		.take(5)
-		.map(|bucket| format!("{} {}", bucket.name, format_seconds_compact(bucket.wall_seconds)))
-		.collect::<Vec<_>>()
-		.join(", ")
-}
-
-fn format_seconds_compact(seconds: i64) -> String {
-	if seconds >= 3_600 {
-		return format!("{}h{}m", seconds / 3_600, (seconds % 3_600) / 60);
-	}
-	if seconds >= 60 {
-		return format!("{}m{}s", seconds / 60, seconds % 60);
-	}
-
-	format!("{seconds}s")
-}
-
-fn format_count_compact(count: i64) -> String {
-	if count >= 1_000_000 {
-		return format!("{:.2}M", count as f64 / 1_000_000.0);
-	}
-	if count >= 1_000 {
-		return format!("{:.1}k", count as f64 / 1_000.0);
-	}
-
-	count.to_string()
-}
-
-fn format_bytes_compact(bytes: i64) -> String {
-	if bytes >= 1_048_576 {
-		return format!("{:.1}MiB", bytes as f64 / 1_048_576.0);
-	}
-	if bytes >= 1_024 {
-		return format!("{:.1}KiB", bytes as f64 / 1_024.0);
-	}
-
-	format!("{bytes}B")
-}
-
-fn append_rendered_history_lane(output: &mut String, lane: &OperatorHistoryLaneStatus) {
-	output.push_str(&format!(
-		"- issue: {}\n  project_id: {}\n  issue_id: {}\n  issue_identifier: {}\n  title: {}\n  attempts: {}\n  ledger_status: {}\n  outcome: {}\n",
-		lane.issue_key,
-		lane.project_id,
-		lane.issue_id,
-		lane.issue_identifier.as_deref().unwrap_or("none"),
-		lane.title.as_deref().unwrap_or("none"),
-		lane.attempt_count,
-		lane.ledger_outcome.ledger_status,
-		lane.ledger_outcome.final_outcome
-	));
-
-	append_rendered_history_ledger_outcome(output, &lane.ledger_outcome);
-
-	output.push_str(&format!(
-		"  lifecycle_metrics: {}\n",
-		render_lane_lifecycle_metrics(&lane.lifecycle_metrics)
-	));
-
-	if history_ledger_outcome_has_records(&lane.ledger_outcome) {
-		output.push_str(&format!(
-			"  local_attempts: {}\n  latest_run_id: {}\n",
-			lane.attempt_count, lane.latest_run.run_id
-		));
-	} else {
-		append_rendered_run(output, &lane.latest_run);
-	}
-	if lane.lifecycle_metrics.phases.is_empty() {
-		return;
-	}
-
-	output.push_str("  lifecycle_bucket_breakdown:\n");
-
-	for phase in &lane.lifecycle_metrics.phases {
-		output.push_str(&format!(
-			"    - lifecycle_bucket: {} lifecycle_bucket_key: {} attempts: {} sources: recorded={} recovered={} current_snapshot={} captured: {}/{} protocol_events: {} child_events: {} wall: {} tool_calls: {} input_tokens: {} output_tokens: {}\n",
-			phase.label,
-			phase.phase,
-			phase.attempt_count,
-			phase.recorded_attempt_count,
-			phase.recovered_attempt_count,
-			phase.current_snapshot_attempt_count,
-			phase.captured_attempt_count,
-			phase.attempt_count,
-			phase.protocol_event_count,
-			phase.child_event_count,
-			format_seconds_compact(phase.wall_seconds),
-			phase.tool_call_count,
-			phase.input_tokens_cumulative,
-			phase.output_tokens_cumulative,
-		));
-	}
-}
-
-fn render_lane_lifecycle_metrics(metrics: &OperatorLaneLifecycleMetrics) -> String {
-	format!(
-		"attempts={}; sources=recorded:{},recovered:{},current_snapshot:{}; captured={}/{}; missing={}; protocol_events={}; child_events={}; wall={}; tool_calls={}; input_tokens={}; output_tokens={}",
-		metrics.attempt_count,
-		metrics.recorded_attempt_count,
-		metrics.recovered_attempt_count,
-		metrics.current_snapshot_attempt_count,
-		metrics.captured_attempt_count,
-		metrics.attempt_count,
-		metrics.missing_attempt_count,
-		metrics.protocol_event_count,
-		metrics.child_event_count,
-		format_seconds_compact(metrics.wall_seconds),
-		metrics.tool_call_count,
-		metrics.input_tokens_cumulative,
-		metrics.output_tokens_cumulative,
-	)
-}
-
-fn render_lane_lifecycle_evidence(metrics: &OperatorLaneLifecycleMetrics) -> String {
-	if metrics.attempt_evidence.is_empty() && metrics.recovery_gaps.is_empty() {
-		return String::from("none");
-	}
-
-	let mut lines = metrics
-		.attempt_evidence
-		.iter()
-		.map(|attempt| {
-			let evidence = if attempt.evidence.is_empty() {
-				String::from("none")
-			} else {
-				attempt.evidence.join(",")
-			};
-			let gaps = if attempt.gaps.is_empty() {
-				String::from("none")
-			} else {
-				attempt.gaps.join(",")
-			};
-
-			format!(
-				"run={} attempt={} phase={} source={} evidence={} gaps={} protocol_events={} child_events={} updated_at={}",
-				attempt.run_id,
-				attempt.attempt_number,
-				attempt.phase,
-				attempt.source,
-				evidence,
-				gaps,
-				attempt.protocol_event_count,
-				attempt.child_event_count,
-				attempt.updated_at
-			)
-		})
-		.collect::<Vec<_>>();
-
-	if !metrics.recovery_gaps.is_empty() {
-		lines.push(format!("aggregate_gaps={}", metrics.recovery_gaps.join(",")));
-	}
-
-	lines.join(" | ")
-}
-
-fn append_rendered_history_ledger_outcome(
-	output: &mut String,
-	outcome: &OperatorHistoryLedgerOutcome,
-) {
-	append_rendered_history_field(output, "event_type", outcome.final_event_type.as_deref());
-	append_rendered_history_field(output, "event_at", outcome.final_event_at.as_deref());
-	append_rendered_history_field(output, "summary", outcome.summary.as_deref());
-	append_rendered_history_field(output, "pr_url", outcome.pr_url.as_deref());
-	append_rendered_history_field(output, "commit_sha", outcome.commit_sha.as_deref());
-	append_rendered_history_field(output, "branch", outcome.branch.as_deref());
-	append_rendered_history_field(output, "closeout_status", outcome.closeout_status.as_deref());
-	append_rendered_history_field(
-		output,
-		"needs_attention_reason",
-		outcome.needs_attention_reason.as_deref(),
-	);
-	append_rendered_history_field(
-		output,
-		"lifecycle_started_at",
-		outcome.lifecycle_started_at.as_deref(),
-	);
-	append_rendered_history_field(
-		output,
-		"lifecycle_finished_at",
-		outcome.lifecycle_finished_at.as_deref(),
-	);
-
-	if let Some(elapsed) = outcome.lifecycle_elapsed_seconds {
-		output.push_str(&format!("  lifecycle_elapsed_seconds: {elapsed}\n"));
-	}
-
-	output.push_str(&format!("  ledger_records: {}\n", outcome.record_count));
-}
-
-fn append_rendered_history_field(output: &mut String, label: &str, value: Option<&str>) {
-	if let Some(value) = value.filter(|value| !value.trim().is_empty()) {
-		output.push_str(&format!("  {label}: {value}\n"));
-	}
-}
-
-fn history_ledger_outcome_has_records(outcome: &OperatorHistoryLedgerOutcome) -> bool {
-	matches!(outcome.ledger_status.as_str(), "present" | "partial")
-}
-
-fn operator_run_phase_readback(run: &OperatorRunStatus) -> &str {
-	if run.run_phase.trim().is_empty() { &run.phase } else { &run.run_phase }
-}
-
-fn append_rendered_run(output: &mut String, run: &OperatorRunStatus) {
-	let (freshness_source, freshness_at) = operator_run_freshness(run);
-	let protocol_event = render_run_protocol_event(run);
-	let thread_id = run.thread_id.as_deref().unwrap_or("none");
-	let turn_id = run.turn_id.as_deref().unwrap_or("none");
-	let thread_status = run.thread_status.as_deref().unwrap_or("none");
-	let thread_active_flags = render_run_thread_active_flags(run);
-	let idle_for_seconds =
-		run.idle_for_seconds.map_or_else(|| String::from("none"), |value| value.to_string());
-	let protocol_idle_for_seconds = run
-		.protocol_idle_for_seconds
-		.map_or_else(|| String::from("none"), |value| value.to_string());
-	let branch_name = run.branch_name.as_deref().unwrap_or("none");
-	let worktree_path = run.worktree_path.as_deref().unwrap_or("none");
-	let queue_lease = operator_run_queue_lease_summary(run);
-	let child_agent_activity =
-		render_child_agent_activity_summary(run.child_agent_activity.as_ref());
-	let context_pressure = render_child_agent_context_pressure(run.child_agent_activity.as_ref());
-	let protocol_activity = render_protocol_activity_summary(run.protocol_activity.as_ref());
-	let account = render_account_summary(run.account.as_ref());
-	let accounts = render_accounts_summary(&run.accounts);
-	let private_evidence = render_private_evidence_reference(run);
-	let loop_status = render_loop_status_summary(run.loop_status.as_ref());
-	let loop_autonomy_signals = render_loop_autonomy_signals_summary(run.loop_status.as_ref());
-	let loop_review = render_loop_review_summary(run.loop_status.as_ref());
-	let loop_architecture_recovery =
-		render_loop_architecture_recovery_summary(run.loop_status.as_ref());
-	let loop_boundary = render_loop_boundary_summary(run.loop_status.as_ref());
-	let control_capability = render_control_capability_summary(run.control_capability.as_ref());
-	let continuation_recovery =
-		render_continuation_recovery_summary(run.continuation_recovery.as_ref());
-	let phase_acceptance = render_phase_acceptance_summary(run.phase_acceptance.as_ref());
-
-	output.push_str(&format!(
-		"- run_id: {}\n  project_id: {}\n  issue_id: {}\n  issue_identifier: {}\n  title: {}\n  attempt: {}\n  status: {}\n  attempt_status: {}\n  status_projection_reason: {}\n  ownership_state: {}\n  liveness_state: {}\n  policy_state: {}\n  terminalization_state: {}\n  lane_control_next_action: {}\n  lane_control_conditions: {}\n  run_phase: {}\n  wait_reason: {}\n  current_operation: {}\n  active_goal_phase: {}\n  public_progress_phase: {}\n  run_lease: {}\n  queue_lease_state: {}\n  queue_lease: {}\n  execution_liveness: {}\n  has_fresh_execution: {}\n  counts_as_running: {}\n  needs_attention: {}\n  freshness_at: {}\n  freshness_source: {}\n  timing: run_idle={} protocol_idle={} last_progress={} protocol_event={} events={}\n  account: {}\n  accounts: {}\n  child_agent_activity: {}\n  protocol_activity: {}\n  context_pressure: {}\n  lifecycle_metrics: {}\n  lifecycle_evidence: {}\n  private_evidence: {}\n  loop_status: {}\n  loop_autonomy_signals: {}\n  loop_review: {}\n  loop_architecture_recovery: {}\n  loop_boundary: {}\n  control_capability: {}\n  thread_id: {}\n  turn_id: {}\n  thread_status: {}\n  thread_active_flags: {}\n  interactive_requested: {}\n  continuation_pending: {}\n  continuation_recovery: {}\n  phase_acceptance: {}\n  branch: {}\n  worktree_path: {}\n  updated_at: {}\n  last_run_activity_at: {}\n  last_protocol_activity_at: {}\n  last_progress_at: {}\n  idle_for_seconds: {}\n  protocol_idle_for_seconds: {}\n  suspected_stall: {}\n  progress_diagnostic: {}\n  process_id: {}\n  process_alive: {}\n  process_liveness_reason: {}\n  retry_kind: {}\n  next_retry_at: {}\n  effective_model: {}\n  effective_model_provider: {}\n  effective_cwd: {}\n  effective_approval_policy: {}\n  effective_approvals_reviewer: {}\n  effective_sandbox_mode: {}\n  protocol_event: {}\n  event_count: {}\n",
-		run.run_id,
-		run.project_id,
-		run.issue_id,
-		run.issue_identifier.as_deref().unwrap_or("none"),
-		run.title.as_deref().unwrap_or("none"),
-		run.attempt_number,
-		run.status,
-		run.attempt_status,
-		run.status_projection_reason.as_deref().unwrap_or("none"),
-		run.ownership_state,
-		run.liveness_state,
-		run.policy_state,
-		run.terminalization_state,
-		run.lane_control_next_action,
-		render_lane_control_conditions(run),
-		operator_run_phase_readback(run),
-		run.wait_reason.as_deref().unwrap_or("none"),
-		run.current_operation,
-		run.active_goal_phase.as_deref().unwrap_or("none"),
-		run.public_progress_phase.as_deref().unwrap_or("none"),
-		if run.run_lease { "yes" } else { "no" },
-		run.queue_lease_state,
-		queue_lease,
-		run.execution_liveness,
-		if run.has_fresh_execution { "yes" } else { "no" },
-		if run.counts_as_running { "yes" } else { "no" },
-		if run.needs_attention { "yes" } else { "no" },
-		freshness_at,
-		freshness_source,
-		idle_for_seconds,
-		protocol_idle_for_seconds,
-		run.last_progress_at.as_deref().unwrap_or("none"),
-		protocol_event,
-		run.event_count,
-		account,
-		accounts,
-		child_agent_activity,
-		protocol_activity,
-		context_pressure,
-		render_lane_lifecycle_metrics(&run.lifecycle_metrics),
-		render_lane_lifecycle_evidence(&run.lifecycle_metrics),
-		private_evidence,
-		loop_status,
-		loop_autonomy_signals,
-		loop_review,
-		loop_architecture_recovery,
-		loop_boundary,
-		control_capability,
-		thread_id,
-		turn_id,
-		thread_status,
-		thread_active_flags,
-		if run.interactive_requested { "yes" } else { "no" },
-		if run.continuation_pending { "yes" } else { "no" },
-		continuation_recovery,
-		phase_acceptance,
-		branch_name,
-		worktree_path,
-		run.updated_at,
-		run.last_run_activity_at.as_deref().unwrap_or("none"),
-		run.last_protocol_activity_at.as_deref().unwrap_or("none"),
-		run.last_progress_at.as_deref().unwrap_or("none"),
-		idle_for_seconds,
-		protocol_idle_for_seconds,
-		if run.suspected_stall { "yes" } else { "no" },
-		run.progress_diagnostic.as_deref().unwrap_or("none"),
-		run.process_id.map_or_else(|| String::from("none"), |value| value.to_string()),
-		run.process_alive.map_or_else(
-			|| String::from("none"),
-			|value| if value { String::from("yes") } else { String::from("no") },
-		),
-		run.process_liveness_reason.as_deref().unwrap_or("none"),
-		run.retry_kind.as_deref().unwrap_or("none"),
-		run.next_retry_at.as_deref().unwrap_or("none"),
-		run.effective_model.as_deref().unwrap_or("none"),
-		run.effective_model_provider.as_deref().unwrap_or("none"),
-		run.effective_cwd.as_deref().unwrap_or("none"),
-		run.effective_approval_policy.as_deref().unwrap_or("none"),
-		run.effective_approvals_reviewer.as_deref().unwrap_or("none"),
-		run.effective_sandbox_mode.as_deref().unwrap_or("none"),
-		protocol_event,
-		run.event_count
-	));
-}
-
-fn render_run_protocol_event(run: &OperatorRunStatus) -> String {
-	match (&run.last_event_type, &run.last_event_at) {
-		(Some(event_type), Some(timestamp)) => format!("{event_type} @ {timestamp}"),
-		(Some(event_type), None) => event_type.clone(),
-		(None, Some(timestamp)) => timestamp.clone(),
-		(None, None) => String::from("none"),
-	}
-}
-
-fn render_run_thread_active_flags(run: &OperatorRunStatus) -> String {
-	if run.thread_active_flags.is_empty() {
-		String::from("none")
-	} else {
-		run.thread_active_flags.join(",")
-	}
-}
-
-fn render_lane_control_conditions(run: &OperatorRunStatus) -> String {
-	if run.lane_control_conditions.is_empty() {
-		String::from("none")
-	} else {
-		run.lane_control_conditions.join(",")
-	}
-}
-
-fn render_continuation_recovery_summary(
-	recovery: Option<&OperatorContinuationRecoveryStatus>,
-) -> String {
-	let Some(recovery) = recovery else {
-		return String::from("none");
-	};
-	let message = recovery
-		.source_error_message
-		.as_deref()
-		.map(single_line_status_value)
-		.unwrap_or_else(|| String::from("none"));
-
-	format!(
-		"state={} source_phase={} next_phase={} source_error_class={} source_error_message={} count={}/{} budget_exceeded={} recorded_at={} run_id={} attempt={} next_action={}",
-		recovery.state,
-		recovery.source_phase,
-		recovery.next_phase,
-		recovery.source_error_class,
-		message,
-		recovery.recovery_count,
-		recovery.automatic_continuation_limit,
-		if recovery.budget_exceeded { "yes" } else { "no" },
-		recovery.recorded_at,
-		recovery.run_id,
-		recovery.attempt_number,
-		recovery.next_action,
-	)
-}
-
-fn render_phase_acceptance_summary(acceptance: Option<&OperatorPhaseAcceptanceStatus>) -> String {
-	let Some(acceptance) = acceptance else {
-		return String::from("none");
-	};
-	let surfaces = if acceptance.changed_surfaces.is_empty() {
-		String::from("none")
-	} else {
-		acceptance.changed_surfaces.join(",")
-	};
-
-	format!(
-		"phase={} decision={} reason={} objective_covered={} effective_delta={} surfaces={} non_goal_passed={} validation_passed={} recorded_at={} run_id={} attempt={} next_action={}",
-		acceptance.phase,
-		acceptance.decision,
-		acceptance.reason_code,
-		if acceptance.objective_covered { "yes" } else { "no" },
-		if acceptance.effective_delta_present { "yes" } else { "no" },
-		surfaces,
-		if acceptance.non_goal_passed { "yes" } else { "no" },
-		if acceptance.validation_passed { "yes" } else { "no" },
-		acceptance.recorded_at,
-		acceptance.run_id,
-		acceptance.attempt_number,
-		acceptance.next_action
-	)
-}
-
-fn single_line_status_value(value: &str) -> String {
-	value.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-fn operator_run_queue_lease_summary(run: &OperatorRunStatus) -> String {
-	if run.run_lease {
-		return String::from("held");
-	}
-
-	match run.execution_liveness.as_str() {
-		"process_alive" => String::from("not_held (process_alive keeps lane visible)"),
-		"thread_active" => String::from("not_held (thread_active keeps lane visible)"),
-		"protocol_observed" => String::from("not_held (protocol_observed keeps lane visible)"),
-		"process_stopped" => String::from("not_held (process_stopped needs attention)"),
-		EXECUTION_LIVENESS_PROCESS_IDENTITY_MISMATCH
-			if operator_run_has_recent_app_server_execution(run) =>
-			String::from("not_held (app_server_activity keeps lane visible)"),
-		EXECUTION_LIVENESS_PROCESS_IDENTITY_MISMATCH =>
-			String::from("not_held (process_identity_mismatch needs attention)"),
-		_ => String::from("not_held"),
-	}
-}
-
-fn operator_run_freshness(run: &OperatorRunStatus) -> (&'static str, &str) {
-	if operator_run_counts_as_current_lane(run) {
-		if let Some(timestamp) = run.last_run_activity_at.as_deref() {
-			return ("last_run_activity_at", timestamp);
-		}
-		if let Some(timestamp) = run.last_progress_at.as_deref() {
-			return ("last_progress_at", timestamp);
-		}
-		if let Some(timestamp) = run.last_protocol_activity_at.as_deref() {
-			return ("last_protocol_activity_at", timestamp);
-		}
-
-		return ("none", "none");
-	}
-
-	("updated_at", run.updated_at.as_str())
 }
